+++ src/main/java/exercises/xf/model/Drink.java
@@ -1,13 +1,13 @@
 package exercises.xf.model;
 
-import java.util.Objects;
+import java.math.BigDecimal;
 
 /**
  * Created by guisil on 31/01/2017.
  */
 public class Drink extends Supply {
 
-    Drink(String name, double price, Cuisine cuisine) {
+    Drink(String name, BigDecimal price, Cuisine cuisine) {
         super(name, price, cuisine);
     }
