+++ src/main/java/exercises/xf/model/MainCourse.java
@@ -1,11 +1,13 @@
 package exercises.xf.model;
 
+import java.math.BigDecimal;
+
 /**
  * Created by guisil on 31/01/2017.
  */
 public class MainCourse extends Supply {
 
-    MainCourse(String name, double price, Cuisine cuisine) {
+    MainCourse(String name, BigDecimal price, Cuisine cuisine) {
         super(name, price, cuisine);
 
         if (cuisine == null) {
